refactor: moved sentinal to client/types.py
