+++ changes/1778.fix.md
@@ -1,0 +1,4 @@
+option.click type=bool -> type=OptionalType(BoolExprType)
+moved params.py from client/cli to cli since it is used by admin/cli
+moved Undefined class from client/cli/types.py to cli/types due to dependency from OptionalType in params.py
+refactored test fixture
