+++ src/ai/backend/cli/types.py
@@ -6,9 +6,7 @@ from typing import Dict
 import attr
 
 __all__ = (
-    "Sentinel",
     "Undefined",
-    "sentinel",
     "undefined",
 )
 
@@ -26,18 +24,6 @@ class ExitCode(enum.IntEnum):
     INVALID_ARGUMENT = 4  # invalid argument while it's not UsageError
 
 
-class Sentinel(enum.Enum):
-    """
-    A special type to represent a special value to indicate closing/shutdown of queues.
-    """
-
-    TOKEN = 0
-
-    def __bool__(self) -> bool:
-        # It should be evaluated as False when used as a boolean expr.
-        return False
-
-
 class Undefined(enum.Enum):
     """
     A special type to represent an undefined value.
@@ -50,5 +36,4 @@ class Undefined(enum.Enum):
         return False
 
 
-sentinel = Sentinel.TOKEN
 undefined = Undefined.TOKEN
